Refactoring and add ChartGame

+++ app/src/main/java/com/orange_infinity/rileywheelsimulator/entities_layer/ItemBox.kt
@@ -4,5 +4,5 @@ import com.orange_infinity.rileywheelsimulator.entities_layer.items.Item
 
 data class ItemBox(val item: Item, var count: Int) : Comparable<ItemBox> {
 
-    override fun compareTo(other: ItemBox): Int = other.item.getCost() - item.getCost()
+    override fun compareTo(other: ItemBox): Int = (other.item.getCost() - item.getCost()).toInt()
 }
