+++ app/src/main/java/com/orange_infinity/rileywheelsimulator/entities_layer/items/Arcana.kt
@@ -7,15 +7,15 @@ import java.util.*
 import android.os.Parcel
 import android.os.Parcelable
 
-enum class Arcana(private val itemName: String, private val cost: Int, val hero: Heroes, val rarity: Rarity) : Item {
-    BladesOfVothDomosh("Blades of Voth Domosh", cost = 10, hero = Heroes.LegionCommander, rarity = Rarity.ARCANA),
-    DemonEater("Demon Eater", cost = 7, hero = Heroes.ShadowFiend, rarity = Rarity.ARCANA),
-    FierySoulOfTheSlayer("Fiery Soul of the Slayer", cost = 14, hero = Heroes.Lina, rarity = Rarity.ARCANA),
-    FractalHornsOfInnerAbysm("Fractal Horns of Inner Abysm", cost = 6, hero = Heroes.Terrorblade, rarity = Rarity.ARCANA),
-    FrostAvalanche("Frost Avalanche", cost = 24, hero = Heroes.CrystalMaiden, rarity = Rarity.ARCANA),
-    ManifoldParadox("Manifold Paradox", cost = 7, hero = Heroes.PhantomAssassin, rarity = Rarity.ARCANA),
-    SwineOfYheSunkenGalley("Swine of the Sunken Galley", cost = 8, hero = Heroes.Techies, rarity = Rarity.ARCANA),
-    TempestHelmOfTheThundergod("Tempest Helm of the Thundergod", cost = 10, hero = Heroes.Zeus, rarity = Rarity.ARCANA);
+enum class Arcana(private val itemName: String, private val cost: Float, val hero: Heroes, val rarity: Rarity) : Item {
+    BladesOfVothDomosh("Blades of Voth Domosh", cost = 10f, hero = Heroes.LegionCommander, rarity = Rarity.ARCANA),
+    DemonEater("Demon Eater", cost = 7f, hero = Heroes.ShadowFiend, rarity = Rarity.ARCANA),
+    FierySoulOfTheSlayer("Fiery Soul of the Slayer", cost = 14f, hero = Heroes.Lina, rarity = Rarity.ARCANA),
+    FractalHornsOfInnerAbysm("Fractal Horns of Inner Abysm", cost = 6f, hero = Heroes.Terrorblade, rarity = Rarity.ARCANA),
+    FrostAvalanche("Frost Avalanche", cost = 24f, hero = Heroes.CrystalMaiden, rarity = Rarity.ARCANA),
+    ManifoldParadox("Manifold Paradox", cost = 7f, hero = Heroes.PhantomAssassin, rarity = Rarity.ARCANA),
+    SwineOfYheSunkenGalley("Swine of the Sunken Galley", cost = 8f, hero = Heroes.Techies, rarity = Rarity.ARCANA),
+    TempestHelmOfTheThundergod("Tempest Helm of the Thundergod", cost = 10f, hero = Heroes.Zeus, rarity = Rarity.ARCANA);
 
 //    private val arcanaValues: Array<Arcana> = values()
     private val random = Random()
@@ -31,7 +31,7 @@ enum class Arcana(private val itemName: String, private val cost: Int, val hero:
 
     override fun getItemName(): String = itemName
 
-    override fun getCost(): Int = cost
+    override fun getCost(): Float = cost
 
     override fun getRarity(): String = rarity.toString()
 }
