+++ app/src/main/java/com/orange_infinity/rileywheelsimulator/entities_layer/items/Courier.kt
@@ -6,31 +6,31 @@ import java.util.*
 
 enum class Courier(
     private val itemName: String,
-    private val cost: Int,
+    private val cost: Float,
     private val rarity: Rarity = Rarity.MYTHICAL
 ) : Item {
-    CaptainBamboo("Captain Bamboo", 12),
-    YonexsRage("Yonex's Rage", 34, rarity = Rarity.LEGENDARY),
-    Shagbark("Shagbark", 15, rarity = Rarity.LEGENDARY),
-    NimbleBen("Nimble Ben", 2, rarity = Rarity.LEGENDARY),
-    KupuTheMetamorpher("Kupu the Metamorpher", 1),
-    TheLlamaLlama("The Llama Llama", 55),
-    Itsy("Itsy", 32, rarity = Rarity.LEGENDARY),
-    Mok("Mok", 14, rarity = Rarity.LEGENDARY),
-    BlottoAndStick("Blotto and Stick", 6, rarity = Rarity.LEGENDARY),
-    Tinkbot("Tinkbot", 67),
-    AlphidOfLecaciida("Alphid of Lecaciida", 76),
-    WaldiTheFaithful("Waldi the Faithful", 69, rarity = Rarity.LEGENDARY),
-    ArnabusTheFairyRabbit("Arnabus the Fairy Rabbit", 10),
-    Deathripper("Deathripper", 12),
-    CocoTheCourageous("Coco the Courageous", 13),
-    ToryTheSkyGuardian("Tory the Sky Guardian", 31),
-    Throe("Throe", 41),
-    ClucklesTheBrave("Cluckles the Brave", 58, rarity = Rarity.LEGENDARY),
-    Butch("Butch", 21, rarity = Rarity.LEGENDARY),
-    RamnaughtOfUnderwool("Ramnaught of Underwool", 20),
-    PorcinePrincessPenelope("Porcine Princess Penelope", 31),
-    PrismaticDrake("Prismatic Drake", 32);
+    CaptainBamboo("Captain Bamboo", 12f),
+    YonexsRage("Yonex's Rage", 34f, rarity = Rarity.LEGENDARY),
+    Shagbark("Shagbark", 15f, rarity = Rarity.LEGENDARY),
+    NimbleBen("Nimble Ben", 2f, rarity = Rarity.LEGENDARY),
+    KupuTheMetamorpher("Kupu the Metamorpher", 1f),
+    TheLlamaLlama("The Llama Llama", 55f),
+    Itsy("Itsy", 32f, rarity = Rarity.LEGENDARY),
+    Mok("Mok", 14f, rarity = Rarity.LEGENDARY),
+    BlottoAndStick("Blotto and Stick", 6f, rarity = Rarity.LEGENDARY),
+    Tinkbot("Tinkbot", 67f),
+    AlphidOfLecaciida("Alphid of Lecaciida", 76f),
+    WaldiTheFaithful("Waldi the Faithful", 69f, rarity = Rarity.LEGENDARY),
+    ArnabusTheFairyRabbit("Arnabus the Fairy Rabbit", 10f),
+    Deathripper("Deathripper", 12f),
+    CocoTheCourageous("Coco the Courageous", 13f),
+    ToryTheSkyGuardian("Tory the Sky Guardian", 31f),
+    Throe("Throe", 41f),
+    ClucklesTheBrave("Cluckles the Brave", 58f, rarity = Rarity.LEGENDARY),
+    Butch("Butch", 21f, rarity = Rarity.LEGENDARY),
+    RamnaughtOfUnderwool("Ramnaught of Underwool", 20f),
+    PorcinePrincessPenelope("Porcine Princess Penelope", 31f),
+    PrismaticDrake("Prismatic Drake", 32f);
 
     //    private val courierValues: Array<Courier> = values()
     private val random = Random()
@@ -46,7 +46,7 @@ enum class Courier(
 
     override fun getItemName(): String = itemName
 
-    override fun getCost(): Int = cost
+    override fun getCost(): Float = cost
 
     override fun getRarity(): String = rarity.toString()
 }
